refactor(orders): extract Group C dialog components per global style rules
- Move CancelOrderForm, ServiceCancelButton, ServiceCancelForm,
  ServiceStatusConfirmForm, and ServiceStatusUpdateButton out of the
  route file into features/orders/components/ (kebab-case, one file per
  component).
- Shared types and STATUS_ACTION_LABELS hoisted to
  order-service-dialog.types.ts.
- Convert the five extracted components from function declarations to
  arrow + named-export + FooProps interface form.
- Flip canCancelOrder and isAdmin refund gates from ternary+null to &&
  (show/hide without an else branch).

No behavior change.
